die basic tracer
